refactored code, still errors when downloading
