--- README.md
@@ -1,16 +1,0 @@
-# diffusion2d
-
-## Instructions for students
-
-Please follow the instructions in [pypi_exercise.md](https://github.com/Simulation-Software-Engineering/Lecture-Material/blob/main/03_building_and_packaging/pypi_exercise.md).
-
-The code used in this exercise is based on [Chapter 7 of the book "Learning Scientific Programming with Python"](https://scipython.com/book/chapter-7-matplotlib/examples/the-two-dimensional-diffusion-equation/).
-
-## Description
-
-## Installing the package
-
-## Running this package
-
-## Citing
-
